split the code into different files and add comments
